pyproject.toml
[project]
name = "property-data-extraction"
version = "0.1.0"
description = "Property Data Extraction System using OCR and RAG"
requires-python = ">=3.9"
dependencies = [
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
    "pypdf2>=3.0.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "ragas>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
